chore: top level function, other review improvements
